+++ R/repgrid-elements.r
@@ -1,6 +1,7 @@
-################################################################
+#----------------------------------------------#
 ### 				basic element operations				 ###
-################################################################
+#----------------------------------------------#
+
 # Function that start with e. operate on the elements only.
 # These functions serve for basic operations on elements.
 # In case a function needs to operate on elements and other
@@ -10,7 +11,7 @@
 # create new functions they will be needed.
 
 ### basic functions:
-# -------------------------
+# ------------------------- #
 # add elements
 # delete elements
 # rename elements (full and abbreviated names)
@@ -29,23 +30,27 @@ getElements <- function(x){
 }
 
 
-#' Retrieve element names of repgrid object. 
+#' Retrieve element names of repgrid object.
 #'
-#' Function for convenience, 
-#' so new users do not have to deal with object slots
-#' as they will typically not have knowledge about R object 
-#' structures (S3, S4).
+#' Function for convenience, so new users do not have to deal with object slots
+#' as they will typically not have knowledge about R object structures (S3, S4).
 #'
-#' @param x         \code{repgrid} object.
-#' @return vector   Vector containing the names of the elements.
-#'
+#' @param x \code{repgrid} object.
+#' @return vector Vector containing the names of the elements.
+#' @section Deprecated functions: \code{getElementNames()},
+#'   \code{getElementNames2()}, and \code{eNames()} have been deprecated.
+#'   Instead use \code{elements()}.
 #' @export
 #' @keywords internal
 #' @author  Mark Heckmann
 #'
-getElementNames <- function(x){
+getElementNames <- function(x)
+{
 	if(!inherits(x, "repgrid")) 							# check if x is repgrid object
 		stop("Object x must be of class 'repgrid'.")	
+  
+  .Deprecated("elements")
+  
 	sapply(x@elements, function(x) x$name)
 }
 eNames <- getElementNames
@@ -58,68 +63,11 @@ eNames <- getElementNames
 
 
 
-#' Get or replace element names 
-#' 
-#' Allows to get and set element names. 
-#' Replaces the older functions \code{getElementNames}, \code{getElementNames2},
-#' and \code{eNames} which are deprecated.
-#' 
-#' @param  x A repgrid object.
-#' @rdname elementNames
-#' @export
-#' @examples 
-#' 
-#' ## get element names
-#' e <- elementNames(x)
-#' e
-#' 
-#' ## replace element names
-#' elementNames(x) <- rev(e)       # reverse all element names
-#' elementNames(x)[1] <- "Hannes"  # replace name of first element
-#' 
-#' # replace names of elements 1 and 3
-#' elementNames(x)[c(1,3)] <- c("element 1", "element 3")   
-#'  
-elementNames <- function(x)
-{
-  # check if x is a repgrid object
-  if (!inherits(x, "repgrid")) 
-    stop("Object x must be of class 'repgrid'.")
-  
-  # get name property from each element
-  sapply(x@elements, function(x) x$name)
-}
-
-
-#' @param value Character vector of element names.
-#' @rdname elementNames
-#' @export
-#' 
-`elementNames<-` <- function(x, position, value) 
-{
-  # check if x is a repgrid object
-  if (!inherits(x, "repgrid")) 
-    stop("Object x must be of class 'repgrid'.")
-  
-  # get element names and replace one or more
-  e.names <- elementNames(x)
-  e.names[position] <- value
-  
-  # replace name propert of each element
-  for (i in seq_along(e.names)) {
-    x@elements[[i]]$name <- e.names[i]
-  }
-  
-  x
-}
 
 
 
 
-
-
-
-#' Retrieves the element names from a \code{repgrid}.
+#' DEPRECATED: Retrieves the element names from a \code{repgrid}.
 #' 
 #' Different features like trimming, indexing and choices of seperators
 #' allow to return the kind of format that is needed.
@@ -152,14 +100,17 @@ elementNames <- function(x)
 getElementNames2 <- function(x, trim=20, index=F, 
                              pre="(", post=") " ){
   if (!inherits(x, "repgrid")) 							
-   	 stop("Object x must be of class 'repgrid'")
+    stop("Object x must be of class 'repgrid'")
+  
+  .Deprecated("elements")
+  
   enames <- getElementNames(x)
   
   # add numeric index in front of elements
   if (index)
     ind <- paste(pre, seq_along(enames), post, sep="") else
-    ind  <- ""
-
+      ind  <- ""
+  
   # trim names if prompted
   if (!is.na(trim)){                          
     enames <- substr(enames, 1, trim)
@@ -167,6 +118,64 @@ getElementNames2 <- function(x, trim=20, index=F,
   
   enames.new <- paste(ind, enames, sep="") 
   enames.new
+}
+
+
+
+
+#' Get or replace element names 
+#' 
+#' Allows to get and set element names. 
+#' Replaces the older functions \code{getElementNames}, \code{getElementNames2},
+#' and \code{eNames} which are deprecated.
+#' 
+#' @param  x A repgrid object.
+#' @rdname elements
+#' @export
+#' @examples 
+#' 
+#' ## get element names
+#' e <- elements(x)
+#' e
+#' 
+#' ## replace element names
+#' elements(x) <- rev(e)       # reverse all element names
+#' elements(x)[1] <- "Hannes"  # replace name of first element
+#' 
+#' # replace names of elements 1 and 3
+#' elements(x)[c(1,3)] <- c("element 1", "element 3")   
+#'  
+elements <- function(x)
+{
+  # check if x is a repgrid object
+  if (!inherits(x, "repgrid")) 
+    stop("Object x must be of class 'repgrid'.")
+  
+  # get name property from each element
+  sapply(x@elements, function(x) x$name)
+}
+
+
+#' @param value Character vector of element names.
+#' @rdname elements
+#' @export
+#' 
+`elements<-` <- function(x, position, value) 
+{
+  # check if x is a repgrid object
+  if (!inherits(x, "repgrid")) 
+    stop("Object x must be of class 'repgrid'.")
+  
+  # get element names and replace one or more
+  e.names <- elements(x)
+  e.names[position] <- value
+  
+  # replace name propert of each element
+  for (i in seq_along(e.names)) {
+    x@elements[[i]]$name <- e.names[i]
+  }
+  
+  x
 }